Add files via upload
In this project, you will write a program that prompts a user to sign up for a website. The user must come up with a username and password, then log in using the username and password.

There are requirements for both the username and password. If either one doesn’t meet the requirements, print an appropriate error message and ask the username for a new username and password.

If they are both valid, store them as variables. Then ask the user to log in using the username and password they chose.